+++ frontend/src/components/Sign/SignTabitems.vue
@@ -1,23 +1,23 @@
 <template>
   <v-tabs-items v-model="childTab">
     <v-tab-item value="signin">
-      <sign-container-tab-items-sign-in />
+      <sign-tabitems-sign-in />
     </v-tab-item>
     <v-tab-item value="signup">
-      <sign-container-tab-items-sign-up />
+      <sign-tabitems-sign-up />
     </v-tab-item>
   </v-tabs-items>
 </template>
 
 <script>
 import { mapGetters, mapMutations } from 'vuex'
-import SignContainerTabItemsSignIn from '@/components/Sign/SignContainerTabItemsSignIn.vue'
-import SignContainerTabItemsSignUp from '@/components/Sign/SignContainerTabItemsSignUp.vue'
+import SignTabitemsSignIn from '@/components/Sign/SignTabitemsSignIn.vue'
+import SignTabitemsSignUp from '@/components/Sign/SignTabitemsSignUp.vue'
 
 export default {
   components: {
-    SignContainerTabItemsSignIn,
-    SignContainerTabItemsSignUp
+    SignTabitemsSignIn,
+    SignTabitemsSignUp
   },
 
   computed: {
