+++ frontend/tests/unit/components/Sign/SignContainerTabItems.spec.js
@@ -1,6 +1,6 @@
 import { shallowMount, createLocalVue } from '@vue/test-utils'
 import Vuex from 'vuex'
-import Component from '@/components/Sign/SignContainerTabItems.vue'
+import Component from '@/components/Sign/SignTabitems.vue'
 
 const localVue = createLocalVue()
 localVue.use(Vuex)
