+++ frontend/tests/unit/components/Sign/SignContainerTabItemsSignInOauth.spec.js
@@ -1,6 +1,6 @@
 import { mount, createLocalVue } from '@vue/test-utils'
 import Vuex from 'vuex'
-import Component from '@/components/Sign/SignContainerTabItemsSignInOauth.vue'
+import Component from '@/components/Sign/SignTabitemsSignInOauth.vue'
 
 const localVue = createLocalVue()
 localVue.use(Vuex)
