+++ frontend/tests/unit/components/Sign/SignContainerTabItemsSignUpOauth.spec.js
@@ -1,6 +1,6 @@
 import { mount, createLocalVue } from '@vue/test-utils'
 import Vuex from 'vuex'
-import Component from '@/components/Sign/SignContainerTabItemsSignUpOauth.vue'
+import Component from '@/components/Sign/SignTabitemsSignUpOauth.vue'
 
 const localVue = createLocalVue()
 localVue.use(Vuex)
